Fix dragging from pathview.
